+++ crates/brk_types/src/quarterindex.rs
@@ -23,11 +23,11 @@ use super::MonthIndex;
     Pco,
     JsonSchema,
 )]
-pub struct QuarterIndex(u16);
+pub struct QuarterIndex(u8);
 
-impl From<u16> for QuarterIndex {
+impl From<u8> for QuarterIndex {
     #[inline]
-    fn from(value: u16) -> Self {
+    fn from(value: u8) -> Self {
         Self(value)
     }
 }
@@ -35,11 +35,11 @@ impl From<u16> for QuarterIndex {
 impl From<usize> for QuarterIndex {
     #[inline]
     fn from(value: usize) -> Self {
-        Self(value as u16)
+        Self(value as u8)
     }
 }
 
-impl From<QuarterIndex> for u16 {
+impl From<QuarterIndex> for u8 {
     #[inline]
     fn from(value: QuarterIndex) -> Self {
         value.0
@@ -57,7 +57,7 @@ impl Add<usize> for QuarterIndex {
     type Output = Self;
 
     fn add(self, rhs: usize) -> Self::Output {
-        Self::from(self.0 + rhs as u16)
+        Self::from(self.0 + rhs as u8)
     }
 }
 
@@ -85,7 +85,7 @@ impl Div<usize> for QuarterIndex {
 impl From<MonthIndex> for QuarterIndex {
     #[inline]
     fn from(value: MonthIndex) -> Self {
-        Self((usize::from(value) / 3) as u16)
+        Self((usize::from(value) / 3) as u8)
     }
 }
